Move underlying type tests to new cast_test.go file.
