+++ bytes_test.go
@@ -19,18 +19,6 @@ func TestBytes(t *testing.T) {
 	})
 }
 
-func TestBytesUnderlyingType(t *testing.T) {
-	t.Parallel()
-	type header []byte
-	codec := lexy.NilsLast(lexy.CastBytes[header]())
-	testCodec(t, codec, []testCase[header]{
-		{"nil", header(nil), []byte{pNilLast}},
-		{"empty", header{}, []byte{pNonNil}},
-		{"[0]", header{0}, []byte{pNonNil, 0x00}},
-		{"[1, 2, 3]", header{1, 2, 3}, []byte{pNonNil, 0x01, 0x02, 0x03}},
-	})
-}
-
 func TestBytesNilsLast(t *testing.T) {
 	t.Parallel()
 	testOrdering(t, lexy.NilsLast(lexy.Bytes()), []testCase[[]byte]{
