+++ map_test.go
@@ -57,13 +57,6 @@ func TestMapInt(t *testing.T) {
 	testBasicMap(t, lexy.MapOf(lexy.String(), lexy.Int32()))
 }
 
-type mStringInt map[string]int32
-
-func TestMapUnderlyingType(t *testing.T) {
-	t.Parallel()
-	testBasicMapWithPrefix(t, pNilLast, lexy.NilsLast(lexy.CastMapOf[mStringInt](lexy.String(), lexy.Int32())))
-}
-
 func TestMapSlice(t *testing.T) {
 	t.Parallel()
 	codec := lexy.MapOf(lexy.String(), lexy.SliceOf(lexy.String()))
