+++ pointer_test.go
@@ -70,15 +70,3 @@ func TestPointerNilsLast(t *testing.T) {
 		{"nil", nil, nil},
 	})
 }
-
-type pInt *int32
-
-func TestPointerUnderlyingType(t *testing.T) {
-	t.Parallel()
-	codec := lexy.NilsLast(lexy.CastPointerTo[pInt](lexy.Int32()))
-	testCodec(t, codec, []testCase[pInt]{
-		{"nil", pInt(nil), []byte{pNilLast}},
-		{"*0", pInt(ptr(int32(0))), []byte{pNonNil, 0x80, 0x00, 0x00, 0x00}},
-		{"*-1", pInt(ptr(int32(-1))), []byte{pNonNil, 0x7F, 0xFF, 0xFF, 0xFF}},
-	})
-}
