+++ slice_test.go
@@ -203,26 +203,6 @@ func TestSliceSliceString(t *testing.T) {
 	})
 }
 
-type sInt []int32
-
-func TestSliceUnderlyingType(t *testing.T) {
-	t.Parallel()
-	codec := lexy.NilsLast(lexy.CastSliceOf[sInt](lexy.Int32()))
-	assert.True(t, codec.RequiresTerminator())
-	testCodec(t, codec, []testCase[sInt]{
-		{"nil", sInt(nil), []byte{pNilLast}},
-		{"empty", sInt([]int32{}), []byte{pNonNil}},
-		{"[0]", sInt([]int32{0}), []byte{pNonNil, 0x80, 0x00, 0x00, 0x00}},
-		{"[-1]", sInt([]int32{-1}), []byte{pNonNil, 0x7F, 0xFF, 0xFF, 0xFF}},
-		{"[0, 1, -1]", sInt([]int32{0, 1, -1}), []byte{
-			pNonNil,
-			0x80, 0x00, 0x00, 0x00,
-			0x80, 0x00, 0x00, 0x01,
-			0x7F, 0xFF, 0xFF, 0xFF,
-		}},
-	})
-}
-
 func TestSliceNilsLast(t *testing.T) {
 	t.Parallel()
 	codec := lexy.SliceOf(lexy.Int32())
